+++ src/main/java/roomescape/auth/dto/AuthRequest.java
@@ -1,4 +1,4 @@
-package roomescape.auth;
+package roomescape.auth.dto;
 
 public record AuthRequest(
         String email,
