+++ src/main/java/roomescape/auth/dto/AuthResponse.java
@@ -1,4 +1,4 @@
-package roomescape.auth;
+package roomescape.auth.dto;
 
 public record AuthResponse(
         String token
