+++ src/main/java/roomescape/auth/dto/MemberDetailResponse.java
@@ -1,6 +1,6 @@
-package roomescape.auth;
+package roomescape.auth.dto;
 
-public record LoginMember(
+public record MemberDetailResponse(
         Long id,
         String name,
         String email,
