+++ src/main/java/roomescape/reservation/ReservationController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.auth.LoginMember;
+import roomescape.auth.domain.LoginMember;
 
 @RestController
 public class ReservationController {
